+++ CHANGELOG.md
@@ -8,3 +8,13 @@
 
 - The `scripts` directory is now included with an install, and the premade scripts can
   be copied to your local `scripts` folder using `devenv --install_scripts`.
+
+### Fixes
+
+- Directories are no longer run as scripts
+
+### Scripts
+
+#### Fixes
+
+- `gitignore` now runs